DO NOT MERGE: Merging platform/tools/tradefederation into qt-dev.
Bug: 143908003

Exclude merging into *-plus-aosp branches, since they already include the change.
Merged-In: I8feb5415027348acdb2c4baf980aaa8a73a17826
Merged-In: I7d9a365472dc082e75cb4e89babc5bd40567517b

+++ src/com/android/tradefed/device/metric/FilePullerDeviceMetricCollector.java
@@ -18,6 +18,7 @@ package com.android.tradefed.device.metric;
 import com.android.tradefed.config.Option;
 import com.android.tradefed.device.DeviceNotAvailableException;
 import com.android.tradefed.device.ITestDevice;
+import com.android.tradefed.device.StubDevice;
 import com.android.tradefed.log.LogUtil.CLog;
 import com.android.tradefed.metrics.proto.MetricMeasurement.Metric;
 import com.android.tradefed.util.FileUtil;
@@ -136,14 +137,19 @@ public abstract class FilePullerDeviceMetricCollector extends BaseDeviceMetricCo
 
     }
 
+
     private Entry<String, File> pullMetricFile(
             String pattern, final Map<String, String> currentMetrics) {
         Pattern p = Pattern.compile(pattern);
         for (Entry<String, String> entry : currentMetrics.entrySet()) {
             if (p.matcher(entry.getKey()).find()) {
                 for (ITestDevice device : getDevices()) {
+                    // Skip StubDevices
+                    if (device.getIDevice() instanceof StubDevice) {
+                        continue;
+                    }
                     try {
-                        File attemptPull = device.pullFile(entry.getValue());
+                        File attemptPull = retrieveFile(device, entry.getValue());
                         if (attemptPull != null) {
                             if (mCleanUp) {
                                 device.deleteFile(entry.getValue());
@@ -166,6 +172,19 @@ public abstract class FilePullerDeviceMetricCollector extends BaseDeviceMetricCo
     }
 
     /**
+     * Pull the file from the specified path in the device.
+     *
+     * @param device which has the file.
+     * @param remoteFilePath location in the device.
+     * @return File retrieved from the given path in the device.
+     * @throws DeviceNotAvailableException
+     */
+    protected File retrieveFile(ITestDevice device, String remoteFilePath)
+            throws DeviceNotAvailableException {
+        return device.pullFile(remoteFilePath);
+    }
+
+    /**
      * Pulls the directory and all its content from the device and save it in the
      * host under the host_tmp folder.
      *
@@ -177,6 +196,10 @@ public abstract class FilePullerDeviceMetricCollector extends BaseDeviceMetricCo
         try {
             File tmpDestDir = FileUtil.createTempDir("host_tmp");
             for (ITestDevice device : getDevices()) {
+                // Skip StubDevices
+                if (device.getIDevice() instanceof StubDevice) {
+                    continue;
+                }
                 try {
                     if (device.pullDir(keyDirectory, tmpDestDir)) {
                         if (mCleanUp) {
